# enhance: Expose Locations() and TestLocations() API