@greatroar 2024-06-02
Do you have any benchmark figures to back up your claim that computing the secondary hashes is expensive? Because the Locations function in bits-and-blooms has to compute a murmurhash, whereas this package allows you to reuse the primary hash. Computing doublehash takes around a nanosecond, which should be about as much time as TestLocations needs to fetch the hash from locs.