@weiliu1031 2024-06-04
> Here's a benchmark that I think reflects your use case:
> 
> ```
> func BenchmarkLocations(b *testing.B) {
> 	// Use small filters with lots of hashes so memory hierarchy effects
> 	// are small and double hashing cost is high.
> 	// These settings correspond to 150 keys at FPR = 1e-4.
> 	filters := make([]*Filter, 10)
> 	for i := range filters {
> 		filters[i] = New(4096, 20)
> 	}
> 
> 	r := rand.New(rand.NewSource(0xb1ab1ab1a))
> 
> 	for i := 0; i < 150; i++ {
> 		for _, f := range filters {
> 			f.Add(r.Uint64())
> 		}
> 	}
> 
> 	k := filters[0].K()
> 
> 	const useLocations = false
> 
> 	b.ResetTimer()
> 
> 	for i := 0; i < b.N; i++ {
> 		x := r.Uint64()
> 		var loc []uint64
> 		if useLocations {
> 			loc = Locations(x, k)
> 		}
> 
> 		for _, f := range filters {
> 			if useLocations {
> 				f.TestLocations(loc)
> 			} else {
> 				f.Has(x)
> 			}
> 		}
> 	}
> }
> ```
> 
> When I run this with useLocations = false vs. useLocations = true, I get
> 
> ```
> goos: linux
> goarch: amd64
> pkg: github.com/greatroar/blobloom
> cpu: Intel(R) Core(TM) i7-3770K CPU @ 3.50GHz
>             │    false    │                 true                 │
>             │   sec/op    │   sec/op     vs base                 │
> Locations-8   143.2n ± 6%   345.1n ± 2%  +141.03% (p=0.000 n=10)
> ```
> 
> So using Locations is more than twice as slow. Much of that is memory allocation, which can be reduced by passing a slice to Locations for it to fill. Doing so makes it almost as fast as recomputing the hashes, but still slower:
> 
> ```
> cpu: Intel(R) Core(TM) i7-3770K CPU @ 3.50GHz
>             │    false    │              noalloc               │
>             │   sec/op    │   sec/op     vs base               │
> Locations-8   143.2n ± 6%   155.2n ± 0%  +8.41% (p=0.000 n=10)
> ```

Thank you very much for creating this benchmark. It clearly demonstrates that caching locations under Blobloom is ineffective, so I will be switching to caching hash results instead. However, for Bits-and-Blooms, caching locations is still valuable. To achieve optimal performance, I will refactor the code to support both Blobloom and Bits-and-Blooms logic. Thanks again for your help.